+++ examples/simple/versions.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     flexbot = {
       source  = "igor-feoktistov/flexbot"
-      version = ">= 1.5.4"
+      version = ">= 1.6.0"
     }
   }
   required_version = ">= 0.13"